Use interface, not type, where possible. Factor out DummyScreen of legacy tests
