+++ magellan-legacy/src/test/java/com/wealthfront/magellan/ScreenTest.java
@@ -3,6 +3,8 @@ package com.wealthfront.magellan;
 import android.app.Activity;
 import android.content.Context;
 
+import com.wealthfront.magellan.test.DummyScreen;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
