+++ magellan-library-test/src/main/java/com/wealthfront/magellan/test/AttachingActivity.kt
@@ -1,13 +1,13 @@
-package com.wealthfront.magellan
+package com.wealthfront.magellan.test
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import com.wealthfront.magellan.core.Step
+import com.wealthfront.magellan.core.Navigable
 import com.wealthfront.magellan.lifecycle.setContentScreen
 
 public class AttachingActivity : ComponentActivity() {
   public companion object {
-    public var step: Step<*>? = null
+    public var step: Navigable? = null
   }
 
   override fun onCreate(savedInstanceState: Bundle?) {
