+++ magellan-library/src/test/java/com/wealthfront/magellan/core/NavigableExtensionsKtTest.kt
@@ -1,7 +1,7 @@
 package com.wealthfront.magellan.core
 
 import com.google.common.truth.Truth.assertThat
-import com.wealthfront.magellan.databinding.MagellanDummyLayoutBinding
+import com.wealthfront.magellan.test.databinding.MagellanDummyLayoutBinding
 import com.wealthfront.magellan.lifecycle.attachFieldToLifecycle
 import org.junit.Test
