+++ magellan-library/src/test/java/com/wealthfront/magellan/lifecycle/ActivityLifecycleAdapterTest.kt
@@ -6,8 +6,8 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.launchActivity
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.wealthfront.magellan.AttachingActivity
-import com.wealthfront.magellan.DummyStep
+import com.wealthfront.magellan.test.AttachingActivity
+import com.wealthfront.magellan.test.DummyStep
 import com.wealthfront.magellan.core.Step
 import org.junit.After
 import org.junit.Before
